Refactor storage commonality into StatefulCluster
