+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/ModelUtils.java
@@ -17,10 +17,7 @@ import io.fabric8.kubernetes.api.model.SecretBuilder;
 import io.fabric8.kubernetes.api.model.apps.StatefulSet;
 import io.strimzi.api.kafka.model.CertificateAuthority;
 import io.strimzi.api.kafka.model.CpuMemory;
-import io.strimzi.api.kafka.model.JbodStorage;
-import io.strimzi.api.kafka.model.PersistentClaimStorage;
 import io.strimzi.api.kafka.model.Resources;
-import io.strimzi.api.kafka.model.Storage;
 import io.strimzi.api.kafka.model.TlsSidecar;
 import io.strimzi.api.kafka.model.TlsSidecarLogLevel;
 import io.strimzi.api.kafka.model.template.PodDisruptionBudgetTemplate;
@@ -280,22 +277,6 @@ public class ModelUtils {
             model.templateImagePullSecrets = pod.getImagePullSecrets();
             model.templateSecurityContext = pod.getSecurityContext();
         }
-    }
-
-    /**
-     * Returns whether the given {@code Storage} instance is a persistent claim one or
-     * a JBOD containing at least one persistent volume.
-     *
-     * @param storage the Storage instance to check
-     */
-    public static boolean containsPersistentStorage(Storage storage) {
-        boolean isPersistentClaimStorage = storage instanceof PersistentClaimStorage;
-
-        if (!isPersistentClaimStorage && storage instanceof JbodStorage) {
-            isPersistentClaimStorage |= ((JbodStorage) storage).getVolumes()
-                    .stream().anyMatch(volume -> volume instanceof PersistentClaimStorage);
-        }
-        return isPersistentClaimStorage;
     }
 
     /**
